+++ examples/Example_6_TextShaping.cpp
@@ -6,8 +6,8 @@ Image GetAtlasAsBitmapImage(Trex::Atlas& atlas)
 {
 	Image atlasImage;
 	atlasImage.data = atlas.GetBitmap().data(); // pointer to the atlas bitmap data
-	atlasImage.width = atlas.GetWidth(); // width of the atlas bitmap
-	atlasImage.height = atlas.GetHeight(); // height of the atlas bitmap
+	atlasImage.width = (int)atlas.GetWidth(); // width of the atlas bitmap
+	atlasImage.height = (int)atlas.GetHeight(); // height of the atlas bitmap
 	atlasImage.mipmaps = 1;
 	atlasImage.format = PIXELFORMAT_UNCOMPRESSED_GRAYSCALE; // atlas bitmap format is always 1 byte per pixel (grayscale)
 	return atlasImage;
@@ -30,8 +30,8 @@ void RenderShapedText(float cursorX, float cursorY, const Trex::ShapedGlyphs& gl
 {
 	for (const auto& glyph : glyphs)
 	{
-		float x = cursorX + glyph.xOffset + glyph.info.bearingX;
-		float y = cursorY + glyph.yOffset - glyph.info.bearingY;
+		float x = cursorX + glyph.xOffset + (float)glyph.info.bearingX;
+		float y = cursorY + glyph.yOffset - (float)glyph.info.bearingY;
 
 		RenderGlyph(x, y, glyph.info, atlasTexture);
